requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
click>=8.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
rich>=13.7.0
pyjson5>=2.0.0
